MDL-54945 workshop: Introduce workshop::check_group_membership() method
This is just a refactored code block present at two different places. We
will need the same logic in the portfolio caller class soit made sense
to implement it as a new method, rather than make another copy&paste.

As a side effect, the logic is now properly unit tested.