MDL-58543 upgrade: Replay 4 steps missed for some combinations
Moved 4 steps (2017033100.01-04) that were introduced after master
version was 2017040300 and were not executed always.

They have been moved to 2017040700.01-04, just on top of current
weekly (2017040700.00), so they will be executed by everybody.

As far as the 4 fields are XMLDB ones they are executed conditionally
so it's safe to replay them (for people that already had them run).