MDL-58681 behat: Split checkbox and advcheckbox tests
MDL-58267 converted "Display description on course page" to advcheckbox
(see ca2f9bb23f1ed003d30724558356ff349c4192ee).

Advanced checkboxes cannot be tested without real browser because Goutte
does not support the hidden+checkbox duality (MDL-55386).

So this patch does:

- Convert the NON-JS checks to use the "Force format" field (normal checkbox).
- Move the "Display description on course page" (adv. checkbox) to the JS scenario.
- Complete the 2 groups of checks above so the two ways to tick and untick are covered:
    - I set the field "xxxx" to "yyyy"
    - I set the following fields to these values:
- Complete the 2 groups of checks above so the two ways to verify the values are covered:
    - the field "xxxx" [matches|does not match] value "yyyy"
    - the following fields [match|do not match] these values:

That's all! The scenario is a bit spaghetti but it's way off the scope
of the issue to fix that.