MDL-58777 core: Use component_callback in refresh events task
Previously we did not require the lib.php files for each module
in this ad-hoc task, which meant that the *_refresh_events functions
would only be called if some lucky sequence of events led to the
lib.php file for the module being required.

Instead we should use the component_callback function which can handle
that stuff automagically.