MDL-63277 qtype_ddwtos: Re-implement old YUI JS in AMD
In the rewrite is one improvement: drag items now animate when they
are moved, for example out of a drop place and back home. Other than
that, the functionality is unchanged. Also, as far as possible, the
HTML structure has not been changed (althought with changes to the HTML,
further JS improvements would be possible. However, those have been left
for a future issue).