MDL-64066 behat: increase data export sizes, we were hitting them
With MDL-63660 we changed to a bundled moodle.css approach, that
made dowloaded exports size to increase. Although the expected
sizes already were raised in the issue, it seems that some of
them were not enough, so increasing a little bit more.

Side note, we have MDL-63399 about to try to change these tests...