MDL-63915 core_message: prevent drawer overriding index page widget
The module message_drawer_view_conversation.js is written in a way
where it keeps the conversation data as a static module variable.
It also does the same thing for the render function and generates it
when the conversation is shown, and only holds one instance as a static
module variable. This means which ever version of the messaging UI opens
second will override the first render function.