MDL-66253 calendar: Allow event retrival queries to use indexes
Before this change in most cases the queries generated by the API
would not be able to use an index as the subquery would use
conditions that coulde not be satisfied by a signle index.

By changing it to use UNIONs the database will be able to use an
appropriate index for each query in most cases.