MDL-47750 Atto: allow pasting of images
Note: this works if you have image data on the clipboard (e.g. a
screen-grab). It does not work in most other cases, basically because
web browsers are completely unhelpful in the paste/drop events
they raise. See the detailed analysis of my testing in the tracker
issue. (Hopefully this will improve in future.)